separated RenderAction into module
